\name{OCFS_v0.0.1-package}
\alias{OCFS_v0.0.1-package}
\alias{OCFS_v0.0.1}
\docType{package}
\title{
\packageTitle{OCFS_v0.0.1}
}
\description{
\packageDescription{OCFS_v0.0.1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OCFS_v0.0.1}
\packageIndices{OCFS_v0.0.1}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{OCFS_v0.0.1}

Maintainer: \packageMaintainer{OCFS_v0.0.1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
